uint112 -> uint256

+++ contracts/implementation/PoolCommitter.sol
@@ -23,7 +23,7 @@ contract PoolCommitter is IPoolCommitter, Ownable {
     uint128 public latestCommitUnexecuted;
     uint128 public commitIDCounter;
     mapping(uint128 => Commit) public commits;
-    mapping(uint256 => uint112) public shadowPools;
+    mapping(uint256 => uint256) public shadowPools;
 
     address public factory;
 
@@ -32,7 +32,7 @@ contract PoolCommitter is IPoolCommitter, Ownable {
         factory = _factory;
     }
 
-    function commit(CommitType commitType, uint112 amount) external override {
+    function commit(CommitType commitType, uint256 amount) external override {
         require(amount > 0, "Amount must not be zero");
         uint128 currentCommitIDCounter = commitIDCounter;
         commitIDCounter = currentCommitIDCounter + 1;
@@ -161,12 +161,12 @@ contract PoolCommitter is IPoolCommitter, Ownable {
      */
     function executeCommitment(Commit memory _commit) external override onlySelf {
         ILeveragedPool pool = ILeveragedPool(leveragedPool);
-        uint112 shortBalance = pool.shortBalance();
-        uint112 longBalance = pool.longBalance();
+        uint256 shortBalance = pool.shortBalance();
+        uint256 longBalance = pool.longBalance();
         uint256 _commitType = commitTypeToUint(_commit.commitType);
         shadowPools[_commitType] = shadowPools[_commitType] - _commit.amount;
         if (_commit.commitType == CommitType.LongMint) {
-            uint112 mintAmount = PoolSwapLibrary.getMintAmount(
+            uint256 mintAmount = PoolSwapLibrary.getMintAmount(
                 IERC20(pool.poolTokens()[0]).totalSupply(), // long token total supply,
                 _commit.amount, // amount of quote tokens commited to enter
                 longBalance, // total quote tokens in the long pull
@@ -177,11 +177,11 @@ contract PoolCommitter is IPoolCommitter, Ownable {
             // update long and short balances
             pool.setNewPoolBalances(longBalance + _commit.amount, shortBalance);
         } else if (_commit.commitType == CommitType.LongBurn) {
-            uint112 amountOut = PoolSwapLibrary.getAmountOut(
+            uint256 amountOut = PoolSwapLibrary.getAmountOut(
                 PoolSwapLibrary.getRatio(
                     longBalance,
-                    uint112(
-                        uint112(IERC20(pool.poolTokens()[0]).totalSupply()) +
+                    uint256(
+                        uint256(IERC20(pool.poolTokens()[0]).totalSupply()) +
                             shadowPools[commitTypeToUint(CommitType.LongBurn)] +
                             _commit.amount
                     )
@@ -193,7 +193,7 @@ contract PoolCommitter is IPoolCommitter, Ownable {
             pool.setNewPoolBalances(longBalance - amountOut, shortBalance);
             pool.quoteTokenTransfer(_commit.owner, amountOut);
         } else if (_commit.commitType == CommitType.ShortMint) {
-            uint112 mintAmount = PoolSwapLibrary.getMintAmount(
+            uint256 mintAmount = PoolSwapLibrary.getMintAmount(
                 IERC20(pool.poolTokens()[1]).totalSupply(), // short token total supply
                 _commit.amount,
                 shortBalance,
@@ -203,10 +203,10 @@ contract PoolCommitter is IPoolCommitter, Ownable {
             pool.mintTokens(1, mintAmount, _commit.owner);
             pool.setNewPoolBalances(longBalance, shortBalance + _commit.amount);
         } else if (_commit.commitType == CommitType.ShortBurn) {
-            uint112 amountOut = PoolSwapLibrary.getAmountOut(
+            uint256 amountOut = PoolSwapLibrary.getAmountOut(
                 PoolSwapLibrary.getRatio(
                     shortBalance,
-                    uint112(IERC20(pool.poolTokens()[1]).totalSupply()) +
+                    uint256(IERC20(pool.poolTokens()[1]).totalSupply()) +
                         shadowPools[commitTypeToUint(CommitType.ShortBurn)] +
                         _commit.amount
                 ),
